refactor: move sort into a submodule, add median function
