+++ README.md
@@ -3,6 +3,10 @@ This is the repo for RUST implementation of API3's Beacon Server
 
 # Common
 Common package used for all the subsequent chain implementations.
+To run all test
+```
+cargo test --all
+```
 
 # Solana
 Read up on anchors: `https://project-serum.github.io/anchor/`
